Move status enums to enum file
