+++ src/enums/status.ts
@@ -1,0 +1,10 @@
+export enum Status {
+  Canceled = 'Canceled',
+  Complete = 'Complete',
+  Failed = 'Failed',
+  Incomplete = 'Incomplete',
+  Unchecked = 'Unchecked',
+  Modified = 'Modified',
+  Pending = 'Pending',
+  PartialSuccess = 'Partial Success',
+}
